ASCII codes right-justified, event kinds spelled out, dictionary types
An ASCII code now sits right-justified in the integer it stores, padded
with leading NULs: EUR stores as \0EUR, which is simply the integer of
its own letters, and a shorter code sorts below every longer one it
prefixes. Packing, decoding, the currency kernel and the padding
metadata all follow.

EventType rides AsciiInt64, so eight bytes buy explicit spellings --
ORDER, QUOTE, EXECUTED, ISTATE, INSTRMT -- where four forced
abbreviations. Its columns widen to int64 end to end: the message
parser, the category kernel, the transacted ranking and the instrument
batch.

into_arrow_type answers with a cached dictionary type -- the packed
integer indexing the readable codes -- instead of an Arrow extension
type, so nothing registers and every engine reads the column as the
plain integer it is. The extension class and its registries are gone.

Stores, contracts and the enum pages are regenerated to match.

+++ python/src/rekep/fix/rules.py
@@ -331,12 +331,12 @@ class Rules(Convertible):
         if not rows:
             return pyarrow.array([], pyarrow.string())
 
-        event_codes = compute.fill_null(etypes.cast(pyarrow.int32(), safe=False), 0)
+        event_codes = compute.fill_null(etypes.cast(pyarrow.int64(), safe=False), 0)
         market = compute.fill_null(
             compute.is_in(
                 event_codes,
                 value_set=pyarrow.array(
-                    sorted(EventType.ranked_at_least(EventType.INTENT)), pyarrow.int32()
+                    sorted(EventType.ranked_at_least(EventType.INTENT)), pyarrow.int64()
                 ),
             ),
             False,
